fix/registration_name_surname_in_db
