+++ src/bot/handlers/categories.py
@@ -6,7 +6,7 @@ from src.bot.constants import callback_data
 from src.bot.keyboards import get_categories_keyboard, get_open_tasks_and_menu_keyboard
 from src.bot.services.category import CategoryService
 from src.core.logging.utils import logger_decor
-from src.core.services.user import UserService
+from src.bot.services.user import UserService
 
 
 @logger_decor
